refactor src/commands into cli dir
