+++ src/cli/run-print.ts
@@ -1,6 +1,6 @@
 import { createOpenWikiThreadId, runOpenWikiAgent } from "../agent/index.js";
 import { ensureCodeModeRepoSetup } from "../code-mode.js";
-import type { CliCommand } from "../commands.js";
+import type { CliCommand } from "./parse.js";
 import { getErrorMessage } from "../diagnostics.js";
 import { isDebugMode } from "../ui/format.js";
 import { writePrintErrorDiagnostics } from "./error-diagnostics.js";
